# docs/swagger.yaml
definitions:
  inputoutput.LoginInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    type: object
  inputoutput.LoginOutput:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      type:
        example: Bearer
        type: string
    type: object
  inputoutput.ProfessorProfileOutput:
    properties:
      department:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      university:
        type: string
      user_id:
        type: string
    type: object
  inputoutput.RefreshInput:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  inputoutput.RefreshOutput:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      type:
        example: Bearer
        type: string
    type: object
  inputoutput.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  inputoutput.RegisterOutput:
    properties:
      message:
        type: string
      user_id:
        type: string
    type: object
  inputoutput.RegisterProfessorInput:
    properties:
      department:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      university:
        type: string
    type: object
  inputoutput.RegisterStudentInput:
    properties:
      department:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      research_interests:
        type: string
      role:
        type: string
      university:
        type: string
    type: object
  inputoutput.StudentProfileOutput:
    properties:
      department:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      research_interests:
        type: string
      role:
        type: string
      university:
        type: string
      user_id:
        type: string
    type: object
  shared_handler.ErrorResponse:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inputoutput.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.LoginOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access + refresh token
        pair
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inputoutput.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.RefreshOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/auth/register/admin:
    post:
      consumes:
      - application/json
      description: Create a new admin account with email, name, and password
      parameters:
      - description: Admin registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inputoutput.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.RegisterOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      summary: Register admin user
      tags:
      - Auth
  /api/auth/register/professor:
    post:
      consumes:
      - application/json
      description: Create a new professor account with email, name, password, university,
        and department
      parameters:
      - description: Professor registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inputoutput.RegisterProfessorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.RegisterOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      summary: Register professor user
      tags:
      - Auth
  /api/auth/register/student:
    post:
      consumes:
      - application/json
      description: Create a new student account with email, name, password, university,
        department, and research interests
      parameters:
      - description: Student registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inputoutput.RegisterStudentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.RegisterOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      summary: Register student user
      tags:
      - Auth
  /api/professor/profile:
    get:
      description: Returns the authenticated professor's profile derived from JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.ProfessorProfileOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get professor profile
      tags:
      - Professor
  /api/student/profile:
    get:
      description: Returns the authenticated student's profile derived from JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inputoutput.StudentProfileOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get student profile
      tags:
      - Student
  /hello:
    get:
      description: Returns a simple hello world message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: Hello
      summary: Hello World
      tags:
      - Health
swagger: "2.0"